requirements.txt
django== 4.2.10
psycopg2-binary>=2.9.9
six == 1.16.0
django-cors-headers>=4.0.0
python-decouple >= 3.8 
PyJWT >= 2.8.0
whitenoise >= 6.6.0
